backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==4.2.27
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
sendgrid==6.11.0
idna==3.11
PyJWT==2.10.1
redis==7.0.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
